ARM RPi: Add GPIO controller
The Raspberry Pi has a GPIO controller which serves as GPIO pin in/out
detect/select device, but also as pinmux for the pins of the Raspberry Pi.

Recent U-Boot determines its serial port configuration depending on the
pinmux configuration, so we need to populate that with reasonable values.

As of now, pinmux reconfiguration is not supported yet. This will need
to be added when we want to allow dynamic pin reconfiguaration from Linux.

Signed-off-by: Alexander Graf <agraf@suse.de>