applesmc: Ensure debug macro is always compiled
With the current #define logic to enable or disable debugging based on
the preprocessor, we leave out information to the compiler on what
variables are potentially in use.

The real, long term fix would be to move all the debug prints into
breakpoints. But this is a simple quick fix that at least allows us to
make use of variables only inside of debug context.

So change the macro to always contain an fprintf(), albeit with debug
gated by a constant, so the path gets removed by the dead code analysis
step of the compiler.

Signed-off-by: Alexander Graf <agraf@csgraf.de>